refactor: Organize module utilities in dedicated folder
- Move all module utilities to utils/module/ directory
- Rename files for better clarity:
  - moduleRegistry.ts → module/registry.ts
  - moduleConstants.ts → module/constants.ts
  - moduleTabHelper.ts → module/tabHelper.ts
- Update all import paths accordingly
- Improve organization and separation of concerns
